+++ AssemblyBuild/Zenject-usage/IEventRegistrable.cs
@@ -1,0 +1,9 @@
+namespace Zenject
+{
+    public interface IEventRegistrable
+    {
+        void RegisterEvents();
+
+        void UnregisterEvents();
+    }
+}