+++ UnityProject/Assets/Plugins/Zenject/Source/Runtime/Install/ZenjectManagersInstaller.cs
@@ -18,7 +18,7 @@ namespace Zenject
 
         public override void InstallBindings()
         {
-            Container.Bind(typeof(TickableManager), typeof(InitializableManager), typeof(DisposableManager), typeof(LateInitializableManager))
+            Container.Bind(typeof(TickableManager), typeof(InitializableManager), typeof(DisposableManager), typeof(LateInitializableManager), typeof(EventRegistrableManager))
                 .ToSelf().AsSingle().CopyIntoAllSubContainers();
         }
     }